+++ choonchernlim-archetype-webapp-webapp/choonchernlim-archetype-webapp-webapp-war/src/main/frontend/gulp/tasks/optimize-images.js
@@ -2,12 +2,10 @@ var changed = require( 'gulp-changed' );
 var gulp = require( 'gulp' );
 var imageop = require( 'gulp-image-optimization' );
 var config = require( '../config' ).images;
-var browserSync = require( '../util/browser-sync-instance' );
 
 gulp.task( 'optimize-images', function () {
     return gulp.src( config.src )
         .pipe( changed( config.dest ) ) // Ignore unchanged files
         .pipe( imageop() ) // Optimize
-        .pipe( gulp.dest( config.dest ) )
-        .pipe( browserSync.reload( {stream : true} ) );
+        .pipe( gulp.dest( config.dest ) );
 } );