+++ choonchernlim-archetype-webapp-webapp/choonchernlim-archetype-webapp-webapp-war/src/main/frontend/gulp/tasks/sass.js
@@ -1,22 +1,18 @@
 var gulp = require( 'gulp' );
-var browserSync = require( '../util/browser-sync-instance' );
 var sass = require( 'gulp-sass' );
 var handleErrors = require( '../util/handle-errors' );
 var config = require( '../config' ).sass;
 var autoprefixer = require( 'gulp-autoprefixer' );
 var rename = require( 'gulp-rename' );
 var minifyCSS = require( 'gulp-minify-css' );
-var changed = require( 'gulp-changed' );
 
 gulp.task( 'sass', function () {
     return gulp.src( config.src )
         .pipe( sass( config.settings ) )
-        .pipe( changed( config.dest ) ) // Ignore unchanged files
         .on( 'error', handleErrors )
-        .pipe( autoprefixer( {browsers : ['last 2 version']} ) )
+        .pipe( autoprefixer( { browsers : ['last 2 version'] } ) )
         .pipe( gulp.dest( config.dest ) )
-        .pipe( minifyCSS( {keepSpecialComments : 0} ) )
-        .pipe( rename( {suffix : '.min'} ) )
-        .pipe( gulp.dest( config.dest ) )
-        .pipe( browserSync.reload( {stream : true} ) );
+        .pipe( minifyCSS( { keepSpecialComments : 0 } ) )
+        .pipe( rename( { suffix : '.min' } ) )
+        .pipe( gulp.dest( config.dest ) );
 } );