+++ choonchernlim-archetype-webapp-webapp/choonchernlim-archetype-webapp-webapp-war/src/main/frontend/gulp/util/browser-sync-instance.js
@@ -1,7 +1,0 @@
-/*
-    Creates a single instance of browser sync.
- */
-
-'use strict';
-
-module.exports = require( 'browser-sync' ).create();